Restructure, remove FhenixClientV2 and move logic to store based sdk
