+++ src/fhenix.ts
@@ -5,6 +5,9 @@
 
 export { FhenixClient, FhenixClientSync } from "./sdk/index.js";
 
+export { fhenixsdk } from "./sdk/sdk-permitV2.js";
+export { PermitV2 } from "./sdk/permitV2.js";
+
 export {
   ContractPermits,
   InstanceParams,
