+++ src/sdk/sdk-permitV2.ts
@@ -1,30 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { createStore } from "zustand/vanilla";
-import { persist } from "zustand/middleware";
-import { TfheCompactPublicKey } from "../sdk/fhe/fhe";
-import { produce } from "immer";
-import {
-  PermissionV2,
-  PermitV2Interface,
-  PermitV2Options,
-  SendFn,
-  SignTypedDataFn,
-} from "./types";
-import { InitFhevm } from "../sdk/init";
-import {
-  FheOpsAddress,
-  MAX_UINT16,
-  MAX_UINT32,
-  MAX_UINT8,
-  PUBLIC_KEY_LENGTH_MIN,
-} from "../sdk/consts";
-import {
-  toABIEncodedUint32,
-  fromHexString,
-  ValidateUintInRange,
-} from "../sdk/utils";
-import * as tfheEncrypt from "../sdk/encrypt.js";
-import { PermitV2 } from "../sdk/permitV2";
+import { MAX_UINT16, MAX_UINT32, MAX_UINT8 } from "./consts.js";
+import { ValidateUintInRange } from "./utils.js";
+import * as tfheEncrypt from "./encrypt.js";
+import { PermitV2 } from "./permitV2.js";
 import {
   EncryptedNumber,
   EncryptionTypes,
@@ -34,7 +12,10 @@ import {
   Result,
   ResultErr,
   ResultOk,
-} from "../sdk/types";
+  PermissionV2,
+  PermitV2Interface,
+  PermitV2Options,
+} from "./types.js";
 import {
   getActivePermit,
   getPermits as getPermitsFromStore,
@@ -42,219 +23,27 @@ import {
   setActivePermitHash,
   setPermit,
   getActivePermitHash,
-} from "./permitsStore";
-import { isString } from "../sdk/validation.js";
-
-type ChainRecord<T> = Record<string, T>;
-type SecurityZoneRecord<T> = Record<number, T>;
-
-type SdkStore = {
-  initialized: boolean;
-  fhevmInitialized: boolean;
-  fheKeysInitialized: boolean;
-
-  securityZones: number[];
-  fheKeys: ChainRecord<SecurityZoneRecord<Uint8Array | undefined>>;
-
-  account?: string;
-  chainId?: string;
-  send?: SendFn;
-  signTypedData?: SignTypedDataFn;
-};
-
-export type InitParams = {
-  account: string;
-  send: SendFn;
-  signTypedData: SignTypedDataFn;
-  securityZones?: number[];
-  ignoreErrors?: boolean;
-};
-
-// Stores generated permits for each user, a hash indicating the active permit for each user, and a list of fheKeys as a cache
-// Can be used to create reactive hooks
-export const fhenixSdkStore = createStore<SdkStore>()(
-  persist(
-    () => ({
-      initialized: false as boolean,
-      fhevmInitialized: false as boolean,
-      fheKeysInitialized: false as boolean,
-
-      securityZones: [0],
-      fheKeys: {},
-    }),
-    { name: "fhenixsdk" },
-  ),
-);
-
-const getFheKey = (
-  chainId: string | undefined,
-  securityZone: number | undefined,
-): TfheCompactPublicKey | undefined => {
-  if (chainId == null || securityZone == null) return undefined;
-
-  const serialized = fhenixSdkStore.getState().fheKeys[chainId]?.[securityZone];
-  if (serialized == null) return undefined;
-
-  return TfheCompactPublicKey.deserialize(serialized);
-};
-
-const getCurrentChainFheKey = (
-  securityZone: number | undefined,
-): TfheCompactPublicKey | undefined => {
-  const state = fhenixSdkStore.getState();
-
-  if (securityZone == null) return undefined;
-  if (state.chainId == null) return undefined;
-
-  const serialized = state.fheKeys[state.chainId]?.[securityZone];
-  if (serialized == null) return undefined;
-
-  return TfheCompactPublicKey.deserialize(serialized);
-};
-
-const setFheKey = (
-  chainId: string | undefined,
-  securityZone: number | undefined,
-  fheKey: TfheCompactPublicKey | undefined,
-) => {
-  if (chainId == null || securityZone == null) return;
-
-  fhenixSdkStore.setState(
-    produce<SdkStore>((state) => {
-      if (state.fheKeys[chainId] == null) state.fheKeys[chainId] = {};
-      state.fheKeys[chainId][securityZone] = fheKey?.serialize();
-    }),
-  );
-};
-
-const getChainIdFromProvider = async (send: SendFn): Promise<string> => {
-  const chainId = (await send("eth_chainId").catch((err: Error) => {
-    throw Error(`Error while requesting chainId from provider: ${err}`);
-  })) as string;
-
-  if (isNaN(parseInt(chainId, 16))) {
-    throw new Error(
-      `received non-hex number from chainId request: "${chainId}"`,
-    );
-  }
-
-  return chainId;
-};
-
-/**
- * Retrieves the FHE public key from the provider.
- * If the key already exists in the store it is returned, else it is fetched, stored, and returned
- * @param {SupportedProvider} provider - The provider from which to retrieve the key.
- * @param securityZone - The security zone for which to retrieve the key (default 0).
- * @returns {Promise<TfheCompactPublicKey>} - The retrieved public key.
- */
-const fetchFheKey = async (
-  chainId: string | undefined,
-  securityZone: number = 0,
-  send: SendFn | undefined,
-): Promise<TfheCompactPublicKey> => {
-  const storedKey = getFheKey(chainId, securityZone);
-  if (storedKey != null) return storedKey;
-
-  const funcSig = "0x1b1b484e"; // cast sig "getNetworkPublicKey(int32)"
-  const callData = funcSig + toABIEncodedUint32(securityZone);
-
-  const callParams = [{ to: FheOpsAddress, data: callData }, "latest"];
-
-  const publicKey = await send?.("eth_call", callParams).catch((err: Error) => {
-    throw Error(
-      `Error while requesting network public key from provider for security zone ${securityZone}: ${JSON.stringify(
-        err,
-      )}`,
-    );
-  });
-
-  const chainIdNum: number = parseInt(chainId ?? "NaN", 16);
-  if (isNaN(chainIdNum)) {
-    throw new Error(
-      `received non-hex number from chainId request: "${chainId}"`,
-    );
-  }
-
-  if (typeof publicKey !== "string") {
-    throw new Error("Error using publicKey from provider: expected string");
-  }
-
-  if (publicKey.length < PUBLIC_KEY_LENGTH_MIN) {
-    throw new Error(
-      `Error initializing fhenixjs; got shorter than expected public key: ${publicKey.length}`,
-    );
-  }
-
-  // todo (eshel) verify this
-  // magically know how to decode rlp or w/e returns from the evm json-rpc
-  const buff = fromHexString(publicKey.slice(130));
-
-  try {
-    const key = TfheCompactPublicKey.deserialize(buff);
-    setFheKey(chainId, securityZone, key);
-    return key;
-  } catch (err) {
-    throw new Error(`Error deserializing public key ${err}`);
-  }
-};
+} from "../extensions/store/permits.js";
+import { isString } from "./validation.js";
+import {
+  _sdkStore,
+  _store_getConnectedChainFheKey,
+  _store_getFheKey,
+  _store_initialize,
+  InitParams,
+} from "../extensions/store/sdk.js";
 
 const initialize = async (init: InitParams) => {
-  const {
-    account,
-    send,
-    signTypedData,
-    ignoreErrors = false,
-    securityZones = [0],
-  } = init;
+  const { account, send, signTypedData, securityZones = [0] } = init;
 
-  fhenixSdkStore.setState({ initialized: false });
+  if (account == null) throw new Error("initialize :: missing account");
+  if (send == null) throw new Error("initialize :: missing send function");
+  if (signTypedData == null)
+    throw new Error("initialize :: missing signTypedData function");
+  if (securityZones.length === 0)
+    throw new Error("initialize :: no securityZones provided");
 
-  // FHEVM
-
-  if (!fhenixSdkStore.getState().fhevmInitialized) {
-    await InitFhevm().catch((err: unknown) => {
-      if (ignoreErrors) {
-        return undefined;
-      } else {
-        throw new Error(
-          `Failed to initialize fhenixjs - is the network FHE-enabled? ${err}`,
-        );
-      }
-    });
-    fhenixSdkStore.setState({ fhevmInitialized: true });
-  }
-
-  // PROVIDER
-
-  if (!account || !send || !signTypedData) {
-    throw new Error(
-      "Failed to initialize Fhenix Client - must include account, send function, and signTypedData functions",
-    );
-  }
-  fhenixSdkStore.setState({ account, send, signTypedData });
-
-  const chainId = await getChainIdFromProvider(send);
-
-  if (
-    chainId !== fhenixSdkStore.getState().chainId ||
-    securityZones !== fhenixSdkStore.getState().securityZones
-  )
-    fhenixSdkStore.setState({ fheKeysInitialized: false });
-
-  fhenixSdkStore.setState({ chainId, securityZones });
-
-  // FHE KEYS
-
-  if (securityZones.length === 0) throw new Error("No securityZones provided");
-
-  await Promise.all(
-    securityZones.map((securityZone) =>
-      fetchFheKey(chainId, securityZone, send),
-    ),
-  );
-
-  fhenixSdkStore.setState({ fheKeysInitialized: true, initialized: true });
+  await _store_initialize(init);
 };
 
 // Permit
@@ -269,9 +58,10 @@ const initialize = async (init: InitParams) => {
  * @returns {PermitV2} - Newly created PermitV2
  */
 const createPermit = async (options: PermitV2Options): Promise<PermitV2> => {
-  const state = fhenixSdkStore.getState();
+  const state = _sdkStore.getState();
 
-  if (!state.initialized) throw new Error("createPermit :: not initialized");
+  if (!state.initialized)
+    throw new Error("createPermit :: fhenixsdk not initialized");
   if (state.account == null) throw new Error("createPermit :: missing account");
   if (state.chainId == null) throw new Error("createPermit :: missing chainId");
   if (state.send == null)
@@ -297,9 +87,10 @@ const createPermit = async (options: PermitV2Options): Promise<PermitV2> => {
  * @param {PermitV2Interface} imported - Existing permit
  */
 const importPermit = async (imported: PermitV2Interface) => {
-  const state = fhenixSdkStore.getState();
+  const state = _sdkStore.getState();
 
-  if (!state.initialized) throw new Error("importPermit :: not initialized");
+  if (!state.initialized)
+    throw new Error("importPermit :: fhenixsdk not initialized");
   if (state.account == null) throw new Error("importPermit :: missing account");
 
   const permit = await PermitV2.create(imported);
@@ -321,10 +112,10 @@ const importPermit = async (imported: PermitV2Interface) => {
  * @param {string} hash - The `PermitV2.getHash` of the target permit.
  */
 const selectActivePermit = (hash: string) => {
-  const state = fhenixSdkStore.getState();
+  const state = _sdkStore.getState();
 
   if (!state.initialized)
-    throw new Error("selectActivePermit :: not initialized");
+    throw new Error("selectActivePermit :: fhenixsdk not initialized");
   if (state.account == null)
     throw new Error("selectActivePermit :: missing account");
 
@@ -342,9 +133,10 @@ const selectActivePermit = (hash: string) => {
  * @returns {Result<PermitV2>} - The active permit or permit associated with `hash` as a Result object.
  */
 const getPermit = (hash?: string): Result<PermitV2> => {
-  const state = fhenixSdkStore.getState();
+  const state = _sdkStore.getState();
 
-  if (!state.initialized) throw new Error("getPermit :: not initialized");
+  if (!state.initialized)
+    throw new Error("getPermit :: fhenixsdk not initialized");
   if (state.account == null) throw new Error("getPermit :: missing account");
 
   if (hash == null) {
@@ -356,7 +148,8 @@ const getPermit = (hash?: string): Result<PermitV2> => {
   }
 
   const permit = getPermitFromStore(state.account, hash);
-  if (permit == null) return ResultErr(`permit with hash <${hash}> not found`);
+  if (permit == null)
+    return ResultErr(`getPermit :: permit with hash <${hash}> not found`);
 
   return ResultOk(permit);
 };
@@ -381,9 +174,10 @@ const getPermission = (hash?: string): Result<PermissionV2> => {
  * @returns {Result<Record<string, PermitV2>>} - All stored permits.
  */
 const getAllPermits = (): Result<Record<string, PermitV2>> => {
-  const state = fhenixSdkStore.getState();
+  const state = _sdkStore.getState();
 
-  if (!state.initialized) throw new Error("getAllPermits :: not initialized");
+  if (!state.initialized)
+    throw new Error("getAllPermits :: fhenixsdk not initialized");
   if (state.account == null)
     throw new Error("getAllPermits :: missing account");
 
@@ -404,12 +198,13 @@ const encryptValue = (
   type?: EncryptionTypes,
   securityZone: number = 0,
 ): EncryptedNumber => {
-  const state = fhenixSdkStore.getState();
+  const state = _sdkStore.getState();
 
-  if (!state.initialized) throw new Error("encryptValue :: not initialized");
+  if (!state.initialized)
+    throw new Error("encryptValue :: fhenixsdk not initialized");
   if (state.chainId == null) throw new Error("encryptValue :: missing chainId");
 
-  const fhePublicKey = getFheKey(state.chainId, securityZone);
+  const fhePublicKey = _store_getFheKey(state.chainId, securityZone);
   if (fhePublicKey == null)
     throw new Error(
       `encryptValue :: fheKey not found for chain <${state.chainId}> and securityZone <${securityZone}>`,
@@ -459,7 +254,7 @@ function encrypt<T>(item: T) {
 
   // EncryptableItem
   if (isEncryptableItem(item)) {
-    const fhePublicKey = getCurrentChainFheKey(item.securityZone ?? 0);
+    const fhePublicKey = _store_getConnectedChainFheKey(item.securityZone ?? 0);
     if (fhePublicKey == null)
       throw new Error("encrypt :: fheKey for current chain not found");
 
@@ -483,12 +278,12 @@ function encrypt<T>(item: T) {
   if (typeof item === "object" && item !== null) {
     if (Array.isArray(item)) {
       // Array - recurse
-      return item.map((nestedItem) => this.encrypt(nestedItem));
+      return item.map((nestedItem) => encrypt(nestedItem));
     } else {
       // Object - recurse
       const result: any = {};
       for (const key in item) {
-        result[key] = this.encrypt(item[key]);
+        result[key] = encrypt(item[key]);
       }
       return result;
     }
@@ -505,7 +300,7 @@ function encrypt<T>(item: T) {
  * NOTE: Wrapper around `PermitV2.unseal`
  *
  * @param {string} ciphertext - The encrypted message to unseal.
- * @param {string} account - Users address, defaults to this.account
+ * @param {string} account - Users address, defaults to store.account
  * @param {string} hash - The hash of the permit to use for this operation, defaults to active permitV2 hash
  * @returns bigint - The unsealed message.
  */
@@ -514,10 +309,10 @@ const unsealCiphertext = (
   account?: string,
   hash?: string,
 ): bigint => {
-  const state = fhenixSdkStore.getState();
+  const state = _sdkStore.getState();
 
   if (!state.initialized)
-    throw new Error("unsealCiphertext :: not initialized");
+    throw new Error("unsealCiphertext :: fhenixsdk not initialized");
   if (state.chainId == null)
     throw new Error("unsealCiphertext :: missing chainId");
 
@@ -547,7 +342,7 @@ const unsealCiphertext = (
  * @returns - Recursively unsealed data in the target type, SealedBool -> boolean, SealedAddress -> string, etc.
  */
 function unseal<T>(item: T, account?: string, hash?: string) {
-  const resolvedAccount = account ?? this.account;
+  const resolvedAccount = account ?? _sdkStore.getState().account;
   const resolvedHash = hash ?? getActivePermitHash(resolvedAccount);
   if (resolvedAccount == null || resolvedHash == null) {
     throw new Error(`PermitV2 hash not provided and active PermitV2 not found`);
@@ -566,7 +361,7 @@ function unseal<T>(item: T, account?: string, hash?: string) {
 // Export
 
 export const fhenixsdk = {
-  fhenixSdkStore,
+  store: _sdkStore,
   initialize,
 
   createPermit,
